CSS related code is factored out
