+++ src/DOM.mjs
@@ -14,13 +14,13 @@ import {
   relative2absolute,
   copyIfNewer
 } from './Files.mjs';
-import { fileURLToPath } from 'url';
 import processContents from './ContentProcessor.mjs';
-import { getChapterProcessor, getChapterExtractor } from './Chapters.mjs';
+import { getChapterExtractor } from './Chapters.mjs';
 import getPartExtractor from './Parts.mjs';
 import getPreambleExtractor from './Preamble.mjs';
 import getFilenameMaker from './FilenameMaker.mjs';
 import makeHashTable from './MakeHashTable.mjs';
+import { insertCSS, extractCSS } from './CSS.mjs';
 
 const fsp = fs.promises;
 
@@ -442,54 +442,6 @@ export const copyRelativeFiles = (basefile, outDir) => (dom) => {
     copyIfNewer(toAbsoluteInSrcDir(file))
     (toAbsoluteInOutDir(file)).catch(e => console.log(`    Local file linked from the document is missing: ${toAbsoluteInSrcDir(file)}`)));
 };
-
-export const extractCSS = (outDir) => (rootNode) => {
-  rootNode.find('style').each((i, e) => {
-    const basename = `style${i}.css`;
-    const node = new Cheerio(e);
-    fsp.writeFile(path.join(outDir, basename),
-      new Cheerio(e).contents().text());
-    node.replaceWith(new Cheerio(`<link rel='stylesheet' href='${basename}' type='text/css' />`));
-  });
-  return rootNode;
-};
-
-/**
- *
- * @param {object} config
- */
-export const insertCSS = (config) => (rootNode) => {
-  const { css, outdir } = config;
-  if (!css || css.length == 0) return rootNode;
-  const head = rootNode.find('head');
-  css.forEach(cssFile => head.append(cssLink$(outdir, cssFile)));
-  return rootNode;
-};
-
-/**
- * Returns the link url and also copies the css file into the output directory
- * which causes the side effect.
- *
- * @param {string} outdir path to the output directory
- * @param {string} cssFile path to the css file to include
- */
-const cssLink$ = (outdir, cssFile) => {
-  const basename = path.basename(cssFile);
-  const dest = path.join(outdir, basename);
-  if (cssFile === 'asciidoctor-chunker.css') {
-    import( /* webpackMode: "eager" */
-        './css/asciidoctor-chunker.css') // webpack bundle
-      .then(module => fsp.writeFile(dest, module.default))
-      .catch(e => {
-        const __dirname = path.dirname(fileURLToPath(
-          import.meta.url));
-        const src = path.resolve(__dirname, 'css', 'asciidoctor-chunker.css');
-        copyIfNewer(src)(dest);
-      }); // no bundle, regular file
-  } else
-    copyIfNewer(cssFile)(dest);
-  return `<link rel="stylesheet" href="${basename}" type="text/css" />`;
-}
 
 const insertScript = (rootNode) => {
   rootNode.find('html').append(`
